requirements.txt
requests~=2.25.1
prettytable~=2.1.0
pillow
psutil
git+https://github.com/Rapptz/discord.py
numexpr
discord-ext-ipc
quart